# api/books.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "books.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/books": {
      "get": {
        "operationId": "ListBooks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/toshokanListBooksResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      },
      "post": {
        "operationId": "CreateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/toshokanBook"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/toshokanBook"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/books/{book.book_id}": {
      "patch": {
        "operationId": "UpdateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/toshokanBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book.book_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/toshokanBook"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/books/{book_id}": {
      "get": {
        "operationId": "GetBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/toshokanBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      },
      "delete": {
        "operationId": "DeleteBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    }
  },
  "definitions": {
    "toshokanBook": {
      "type": "object",
      "properties": {
        "book_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "publisher": {
          "type": "string"
        },
        "published_at": {
          "type": "string",
          "format": "date-time"
        },
        "description": {
          "type": "string"
        },
        "isbn13": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        }
      }
    },
    "toshokanListBooksResponse": {
      "type": "object",
      "properties": {
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/toshokanBook"
          }
        }
      }
    }
  }
}